azure-functions/requirements.txt
azure-functions
requests
azure-cosmos
python-dotenv
